SPU2ghz: Major optimization fix to the mixer code. VS2008 was not optimizing the MulShr32su function properly. It's about 3-6x faster now. Also optimized the timestretch manager so that it can take full advantage of VC2008 SSE opts.
* XAudio 2 * Fixed potential crash bug on close (which only started happening after I upgraded my DXSDK).

git-svn-id: http://pcsx2-playground.googlecode.com/svn/trunk@268 a6443dda-0b58-4228-96e9-037be469359c

+++ plugins/spu2ghz/mixer.cpp
@@ -78,7 +78,7 @@ static s32 __forceinline MulShr32( s32 srcval, s32 mulval )
 
 static s32 __forceinline MulShr32su( s32 srcval, u32 mulval )
 {
- 	s64 tmp = ((s64)srcval * mulval );
+ 	s64 tmp = ((s64)srcval * (s32)mulval );
 	return ((s32*)&tmp)[1];
 }
